requirements.txt
httpx>=0.27
tenacity>=8.2
selectolax>=0.3.21
pyyaml>=6.0
pandas>=2.0
numpy>=1.26
orjson>=3.10
tqdm>=4.66
scikit-learn>=1.4
